chore(testing): refactor QMenu, split composable props
